+++ grafana/data/plugins/grafana-exploretraces-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.3.2
+
+
+## 1.3.1
+
+##### Chores
+
+*  Upgrade react-router (#621) (eec1df9e)
+
+##### Bug Fixes
+
+*  Check if usePluginFunctions is available (#616) (99112e44)
+
+
 ## 1.3.0
 
 ##### Chores